release 0.0.8 (#170)
* create avbuffers to be 32-byte aligned in width (#154)

* implementation of cmd-period (#155)

* switch travis to xenial (#157)

* Turn on all warnings, and treat warnings as compilation errors (#158)

* Adds preliminary support for Groups to scinserver (#162)

* Adds group and node ordering to scinsynth (#163)

* include libstdc++ and libgcc_s in AppImage (#167)

* fix apprun based on vanilla xenial testing (#168)

* bump version to 0.0.8 (#169)

+++ src/osc/commands/DefLoadDir.cpp
@@ -32,7 +32,7 @@ void DefLoadDir::processMessage(int argc, lo_arg** argv, const char* types, lo_a
         }
     }
     std::shared_ptr<Address> origin(new Address(address));
-    m_dispatcher->async()->scinthDefLoadDirectory(path, [this, origin, onCompletion](int) {
+    m_dispatcher->async()->scinthDefLoadDirectory(path, [this, origin, onCompletion](size_t) {
         if (onCompletion) {
             m_dispatcher->processMessageFrom(origin->get(), onCompletion);
         }
